Вывод общего размера всех раздач на трекере

+++ src/main/java/ru/students/forumservicediplomproject/entity/Statistics.java
@@ -12,5 +12,5 @@ import java.util.HashMap;
 public class Statistics {
     private HashMap<String, Long> allTrackerPeers;
     private long allTrackerPosts;
-    private long allTrackerSize;
+    private double allTrackerSize;
 }
